Consolidate executable cell rendering and simplify architecture
- Remove unnecessary /content/ directory with duplicate components
- Move all editor rendering back into ExecutableCell (much simpler)
- Remove custom AI keyMap - all cells use same keyboard navigation
- Enable line wrapping for AI cells (markdown language + soft wrap)
- Add SQL language support and proper placeholders
- Fix missing Editor import

All executable cells now use unified rendering with cell-type-specific:
- Language: python (code), sql (sql), markdown (ai)
- Line wrapping: enabled for AI cells, disabled for code/sql
- Placeholders: contextual for each cell type
- Tool approval: preserved for AI cells (temporary)

Eliminates ~150 more lines of near-identical code duplication.
